requirements.txt
# Dependencias principales
requests>=2.31.0
python-dotenv<1.0.0,>=0.13.0
flask>=2.0.0
PyYAML>=6.0.0
marked>=0.7.0

# Utilidades
typing-extensions>=4.8.0
pydantic>=2.5.0

# Logging y debugging
loguru>=0.7.2

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0

# Formateo y linting
black>=23.11.0
flake8>=6.1.0 